Moving the execute tests to Core. Fixing bugs this exposes.
